Renamed folder to `meta`
